+++ src/components/TemplateForm/Form.module.sass
@@ -1,15 +1,20 @@
 .form
-  padding: 60px
-  margin: 0
-  grid-area: main
+  box-sizing: border-box
+
   position: relative
   display: flex
-  box-sizing: border-box
   justify-content: space-between
+
+  grid-area: main
+
+  padding: 60px
+  margin: 0
+
 
 .mainWrapper
   max-width: 60vw
   padding: 0 10px 0 0
+
   display: flex
   flex-direction: column
   gap: 40px
@@ -21,8 +26,9 @@
 
 .title
   margin: 0
-  padding: 0
+
   color: #000
+
   font-family: Inter, sans-serif
   font-size: 38px
   font-weight: 700
@@ -31,119 +37,17 @@
 
 .subtitle
   margin: 0
-  padding: 0
+
   font-family: SF Pro Display, sans-serif
   font-size: 24px
   font-weight: 400
   line-height: 28px
+
   max-width: 1050px
-
-  @media (max-width: 1366px)
-    max-width: 678px
-
-
-.list
-  counter-reset: list-item
-  display: flex
-  flex-direction: column
-  position: relative
-  margin: 0
-  list-style-type: none
-  padding: 0
-
-.item
-  counter-increment: list-item
-  border-left: 1px solid #0D23E9
-  padding-left: 39px
-  display: flex
-  flex-direction: column
-  padding-bottom: 32px
-  gap: 20px
-
-  &:last-child
-    border: none
-    padding-bottom: 0
-
-  &::before
-    background-color: white
-    content: counters(list-item, ".") " "
-    position: absolute
-    border: 1px solid #0D23E9
-    left: -15px
-    border-radius: 50%
-    width: 30px
-    height: 30px
-    color: #5C5C5C
-    text-align: center
-    font-family: Inter, sans-serif
-    font-size: 16px
-    font-weight: 700
-    display: flex
-    justify-content: center
-    align-items: center
-
-
-.itemSuccess
-  counter-increment: list-item
-  border-left: 1px solid #15D434
-  padding-left: 39px
-  display: flex
-  gap: 16px
-  flex-direction: column
-  padding-bottom: 32px
-
-  &::before
-    content: ''
-    display: flex
-    background-position: center
-    background-image: url("../../images/checkmark.svg")
-    background-repeat: no-repeat
-    position: absolute
-    left: -15px
-    border-radius: 50%
-    width: 30px
-    height: 30px
-    background-color: #15D434
-
-
-.title
-  padding: 0
-  margin: 0
-  color: #000
-  font-family: Inter, sans-serif
-  font-size: 28px
-  font-style: normal
-  font-weight: 600
-  line-height: 34px
 
 .wrapper
   display: flex
   flex-wrap: wrap
   gap: 16px
   align-items: center
-
-
-@media screen and (max-width: 1919px)
-  .form
-    padding: 40px
-
-  .title
-    font-size: 28px
-    line-height: 34px
-
-  // .titleWrapper
-  //   width: 739px
-
-  .mainWrapper
-    max-width: 739px
-  .subtitle
-    font-size: 20px
-    line-height: 24px
-  .item
-    padding-bottom: 16px
-
-@media (max-width: 1150px)
-  .mainWrapper
-    max-width: 450px
-
