+++ src/components/TemplateForm/FormInputsList.module.sass
@@ -1,0 +1,12 @@
+.list
+  counter-reset: list-item
+  list-style-type: none
+
+  display: flex
+  flex-direction: column
+
+  position: relative
+
+  margin: 0
+  padding: 0
+
