+++ src/components/TemplateForm/FormInputsListItem.module.sass
@@ -1,0 +1,64 @@
+.item
+  counter-increment: list-item
+  border-left: 1px solid #0D23E9
+  padding-left: 39px
+  display: flex
+  flex-direction: column
+  padding-bottom: 32px
+  gap: 20px
+
+  &:last-child
+    border: none
+    padding-bottom: 0
+
+  &::before
+    background-color: white
+    content: counters(list-item, ".") " "
+    position: absolute
+    border: 1px solid #0D23E9
+    left: -15px
+    border-radius: 50%
+    width: 30px
+    height: 30px
+    color: #5C5C5C
+    text-align: center
+    font-family: Inter, sans-serif
+    font-size: 16px
+    font-weight: 700
+    display: flex
+    justify-content: center
+    align-items: center
+
+
+.itemSuccess
+  counter-increment: list-item
+  border-left: 1px solid #15D434
+  padding-left: 39px
+  display: flex
+  gap: 16px
+  flex-direction: column
+  padding-bottom: 32px
+
+  &::before
+    content: ''
+    display: flex
+    background-position: center
+    background-image: url("../../images/checkmark.svg")
+    background-repeat: no-repeat
+    position: absolute
+    left: -15px
+    border-radius: 50%
+    width: 30px
+    height: 30px
+    background-color: #15D434
+
+
+.title
+  padding: 0
+  margin: 0
+  color: #000
+  font-family: Inter, sans-serif
+  font-size: 28px
+  font-style: normal
+  font-weight: 600
+  line-height: 34px
